requirements.txt
omegaconf==2.3.0
opencv_python==4.12.0.88
PyYAML==6.0.2
setproctitle==1.3.7
torch==2.7.1+cu128
torch_fidelity==0.4.0b0
torchvision==0.22.1
tqdm==4.67.1
transformers==4.55.4
#pytorch_lightning==1.6.5.post0  # optional, only used for vae codebook requantization 
